:Add a matrix distance base class which can be used to implement matrix versions of any distance.
