+++ annfab/distances/__init__.py
@@ -1,3 +1,5 @@
+from .matrix import MatrixDistance
 from .matrix_cosine import MatrixCosineDistance
 
-__all__ = [MatrixCosineDistance, ]
+
+__all__ = [MatrixCosineDistance, MatrixDistance]
